refactored getParticipants
